Removed player pod and added default player

+++ MediaWatermark/Classes/Core/MediaVideoProcessor.swift
@@ -71,7 +71,7 @@ extension MediaProcessor {
         let exportSession = AVAssetExportSession(asset: mixComposition, presetName: AVAssetExportPresetHighestQuality)
         exportSession?.videoComposition = videoComposition
         exportSession?.outputURL = processedUrl
-        exportSession?.outputFileType = AVFileTypeQuickTimeMovie
+        exportSession?.outputFileType = AVFileTypeMPEG4
         
         exportSession?.exportAsynchronously(completionHandler: {
             if exportSession?.status == AVAssetExportSessionStatus.completed {
@@ -109,10 +109,12 @@ extension MediaProcessor {
     }
     
     private func clearTemporaryData(url: URL, completion: ProcessCompletionHandler!) {
-        do {
-            try FileManager.default.removeItem(at: url)
-        } catch {
-            completion(MediaProcessResult(processedUrl: nil, image: nil), error)
+        if (FileManager.default.fileExists(atPath: url.path)) {
+            do {
+                try FileManager.default.removeItem(at: url)
+            } catch {
+                completion(MediaProcessResult(processedUrl: nil, image: nil), error)
+            }
         }
     }
